Version 2.98
git-svn-id: https://daisy-model.googlecode.com/svn/trunk@1124 5f5bf4bb-3343-0410-8278-4fa8eb3319dc

+++ root_system.C
@@ -323,7 +323,7 @@ RootSystem::tick_daily (Treelog& msg, const Soil& soil,
   // Penetration.
   if (IncWRoot > 0)
     {
-      const double i = soil.interval_plus (-Depth);
+      const int i = soil.interval_plus (-Depth);
       double clay_fac = PenClayFac (soil.clay (i));
       double dp = PenPar1 * clay_fac * max (0.0, soil_temperature - PenPar2);
       PotRtDpt = min (PotRtDpt + dp, MaxPen);
